Restructure Exponentiated Gradient into a class with the new API (#45)
* move expgrad into its own directory and eliminate the duplicate tests

* separate files

* add class without implementation

* rename args to X, y, A

* reformat doc

* fit/predict

* move to classes, and fix tests

* add comment about namedtuple

* merge

* make adjustments from Richard's change

* one more DP replacement

* remove quality metric

* logging - switch to logger

* _asdict -> _as_dict

* remove debug

* flake8

* constraints instead of disparity_metric

* aux_data -> group_data

* aux_data --> group_data

* remove unused import

* todos

* merge test files

* flake8
