+++ test/reductions/exponentiated_gradient/test_utilities.py
@@ -1,0 +1,9 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+
+group_data = [str(x) for x in 'AAAAAAA' 'BBBBBBB' 'CCCCCC']
+labels =     [int(x) for x in '0110100' '0010111' '001111']  # noqa: E222
+X1 =         [int(x) for x in '0110101' '0111101' '001011']  # noqa: E222
+X2 =         [int(x) for x in '0000100' '0000011' '111111']  # noqa: E222
+X3 =         [int(x) for x in '1111111' '1111111' '111111']  # noqa: E222
